Split transform.js into multiple files
